[#49]refactor:레이아웃 수정 & sightTab 컴포넌트 분리
